sonic-mgmt: Add support for waitunil port is in stormed state
Add helper function to waituntil a port is in a stormed state
before proceeding. This can be used to replace the sleep
commands.

Some of the pfcwd tests test a fake_storm or a real storm
by using the pfc_gen.py script to send pkts to the port.
It can take a while when a real storm is being generated
for SONiC to poll and report the storm.

Signed-off-by: Aaron Payment <aaronp@arista.com>